Bugfix #2: Correctly fill New Elite Archive
Since acquisition elites can be produced & stored within acq_mes.py,
but not necessarily communicated to the MAP-Loop (this is done to not
violate uniform function interfaces), some of the highest performing
acquisition elites were never stored within the new_elite_archive, that
serves as the sample archive, where new objective solutions are to be
sampled from. This likely led to the issue described in #2, where extremely
high acquisition values could be found, and still not lead to high objective
values.