Gradually increase archive resolution during runtime
As calculation progresses, less acquisition elites
mapping to new bins in the objective archive will
be found - to counter this, it is possible to gradually
increase the acquisition archive resolution during runtime.

It is however recommendable to keep the archive resolution
as low as possible, as this will make archive updates far
cheaper.